Fix some OS X clang warnings.
ext/extractors/csv/csv_extractor.c:155:25: error: implicit conversion loses integer precision: 'long' to 'int' [-Werror,-Wshorten-64-to-32]
src/cursor/cur_join.c:1022:16: error: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long') to 'u_int' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]

+++ src/include/cursor.h
@@ -302,7 +302,7 @@ struct __wt_cursor_join_entry {
 
 	WT_CURSOR_JOIN_ENDPOINT	*ends;		/* reference endpoints */
 	size_t			 ends_allocated;
-	size_t			 ends_next;
+	u_int			 ends_next;
 
 	WT_JOIN_STATS		 stats;		/* Join statistics */
 };
